backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.27.0
opencv-python-headless>=4.9.0
numpy>=1.24.0
shapely>=2.0.0
python-dotenv>=1.0.0
